backend/fastapi/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
python-chess==1.999
stockfish==5.2.0
python-dotenv==1.2.2
google-genai